+++ CHANGELOG.md
@@ -11,6 +11,9 @@ for further information about branching and tagging conventions.
 #### Added
 - Buying atomic swaps
 
+#### Fixed
+- A bug when withdrawal form not closed after successfull operation
+
 ## [1.10.0] - 2019-08-12
 
 ## [1.10.0-rc.7] - 2019-08-09